pyproject.toml
[project]
name = "dataverse-utils"
version = "0.17.1"
description = "Utilities for the Dataverse data respository system"
authors = [
    {name = "Paul Lesack",email = "paul.lesack@ubc.ca"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "beautifulsoup4 (>=4.12.2)",
    "certifi (>=2022.12.7)",
    "chardet (>=5.1.0)",
    "charset-normalizer (>=3.1.0)",
    "dryad2dataverse (>=0.6.1)",
    "idna (>=3.4)",
    "markdown (>=3.2.1)",
    "markdownify (>=0.11.6)",
    "requests (>=2.29.0)",
    "requests-toolbelt (>=1.0.0)",
    "six (>=1.16.0)",
    "soupsieve (>=2.4.1)",
    "urllib3 (>=1.26.15)"
]

keywords =['Harvard Dataverse', 'Dataverse', 'research data management', 'data repository'] 

[tool.poetry]
packages = [{"include" = "dataverse_utils", "from"="src"}]

[tool.poetry.group.server]
optional = true

[tool.poetry.group.server.dependencies]
psycopg2-binary = ">=2.9.10"

[tool.poetry.group.dev.dependencies]
pylint = ">=3.3.6"
pydoc-markdown = ">=4.8.2"
mkdocs = ">=1.6.1"

[project.scripts]
dv_collection_info = 'dataverse_utils.scripts.dv_collection_info:main'
dv_del = 'dataverse_utils.scripts.dv_del:main'
dv_ldc_uploader = 'dataverse_utils.scripts.dv_ldc_uploader:main'
dv_manifest_gen = 'dataverse_utils.scripts.dv_manifest_gen:main'
dv_pg_facet_date = 'dataverse_utils.scripts.dv_pg_facet_date:main [server]'
dv_record_copy = 'dataverse_utils.scripts.dv_record_copy:main'
dv_release = 'dataverse_utils.scripts.dv_release:main'
dv_replace_licence = 'dataverse_utils.scripts.dv_replace_licence:main'
dv_replace_license = 'dataverse_utils.scripts.dv_replace_licence:main'
dv_study_migrator = 'dataverse_utils.scripts.dv_study_migrator:main'
dv_upload_tsv = 'dataverse_utils.scripts.dv_upload_tsv:main'
dv_list_files = 'dataverse_utils.scripts.dv_list_files:main'

[project.urls]
Homepage = 'https://ubc-library-rc.github.io/dataverse_utils'
Repository = 'https://github.com/ubc-library-rc/dataverse_utils.git'
'Issue Tracker' = 'https://github.com/ubc-library-rc/dataverse_utils/issues'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
